+++ CHANGELOG.md
@@ -80,9 +80,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   [GEOD-963](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-963)
 - Make the subtitle in cards for optional on the Trends and Analysis Pages
   [GEOD-961](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-961)
+- Removed all instances of 'UPDATE_TRENDS_REGIONS_POINTS' and associated code
+  [GEOD-975](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-975)
 
 ### Fixed
 - Bug in filtering for custom locations
   [GEOD-925](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-925)
 - Removed the card max-height on the V3 Search Page
   [GEOD-982](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-982)
+- Use the API for Trends by Region and Detail
+  [GEOD-977](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-977)